Rename device selection methods in MeetingManager and remove hooks that expose them
