--- src/hooks/sdk/useSelectAudioOutputDevice.tsx
@@ -1,32 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { DefaultBrowserBehavior } from 'amazon-chime-sdk-js';
-import { useCallback } from 'react';
-import { useLogger } from '../../providers/LoggerProvider';
-
-import { useMeetingManager } from '../../providers/MeetingProvider';
-
-export const useSelectAudioOutputDevice = (): ((
-  deviceId: string
-) => Promise<void>) => {
-  const meetingManager = useMeetingManager();
-  const logger = useLogger();
-
-  const selectDevice = useCallback(
-    async (deviceId: string) => {
-      if (new DefaultBrowserBehavior().supportsSetSinkId()) {
-        await meetingManager.selectAudioOutputDevice(deviceId);
-      } else {
-        logger.error(
-          'AudioOutputControl cannot select audio output device because browser does not support setSinkId operation.'
-        );
-      }
-    },
-    [meetingManager]
-  );
-
-  return selectDevice;
-};
-
-export default useSelectAudioOutputDevice;
