+++ src/index.ts
@@ -140,9 +140,6 @@ export {
 export { useAttendeeStatus } from './hooks/sdk/useAttendeeStatus';
 export { useAttendeeAudioStatus } from './hooks/sdk/useAttendeeAudioStatus';
 export { useSelectVideoQuality } from './hooks/sdk/useSelectVideoQuality';
-export { useSelectVideoInputDevice } from './hooks/sdk/useSelectVideoInputDevice';
-export { useSelectAudioInputDevice } from './hooks/sdk/useSelectAudioInputDevice';
-export { useSelectAudioOutputDevice } from './hooks/sdk/useSelectAudioOutputDevice';
 export { useActiveSpeakersState } from './hooks/sdk/useActiveSpeakersState';
 export { useToggleLocalMute } from './hooks/sdk/useToggleLocalMute';
 export { useMeetingStatus } from './hooks/sdk/useMeetingStatus';
